Version 6.0.2: Fixed error handling in JavaScriptExtensions.ToPromise; added JavaScriptExtensions.ToTask (GitHub Issue #182); added DocumentLoader.MaxCacheSize and DocumentCategory.MaxCacheSize; added code to break event connections on engine disposal (GitHub Issue #183); improved ES6 module support, fixing cycle crash (GitHub Issue #181); added DynamicHostObject (GitHub Issue #180); added BigInt / BigInteger support for V8 (GitHub Issue #176); hardened Assembly.Load call in V8Proxy.cs (GitHub Issue #175); improved V8Update environment isolation to fix some V8 build issues (GitHub Issue #185); updated API documentation. Tested with V8 8.3.110.9.

+++ ClearScript/V8/ClearScriptV8/V8IsolateProxyImpl.h
@@ -63,7 +63,7 @@ namespace V8 {
     private:
 
         SharedPtr<V8Isolate> GetIsolate();
-        static int AdjustConstraint(int value);
+        static size_t AdjustConstraint(int value);
 
         Object^ m_gcLock;
         SharedPtr<V8Isolate>* m_pspIsolate;
